Don't override Rails::Rack::Logger
In Rails 4 Rails::Rack::Logger has a different argument set
and expects the tags for tagged logging. If you try to use
tagged logging following this README it will fail with an
error where it can't find the tags. All the middleware needs
to do is set the log level, no need to inherit from Rails::Rack::Logger